test_id changed t long

+++ src/main/java/com/aad/ws/resource/TestResource.java
@@ -1,8 +1,5 @@
 package com.aad.ws.resource;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -13,7 +10,6 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.aad.ws.dto.UserTest;
 import com.aad.ws.dto.UserTestCollection;
 import com.aad.ws.service.TestsService;
 
@@ -29,32 +25,8 @@ public class TestResource {
 	@GET
 	@Path("{testID}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public UserTestCollection getTestDetails(final @PathParam("testID") int id){
-		//TODO: uncomment when database done
+	public UserTestCollection getTestDetails(final @PathParam("testID") long id){
 		UserTestCollection tests = testsService.getTestDetails(id); 
-		
-		//remove it!
-//		UserTestCollection tests = new UserTestCollection();
-//		List<UserTest> testList = new ArrayList<UserTest>();
-//		
-//		
-//		UserTest test1 = new UserTest();
-//		test1.setScore(64);
-//		test1.setTestId(12);
-//		test1.setTime(100);
-//		test1.setQuesAttented(2);
-//		
-//		UserTest test2 = new UserTest();
-//		test2.setScore(59.9);
-//		test2.setTestId(13);
-//		test2.setTime(101);
-//		test2.setQuesAttented(5);
-//		
-//		testList.add(test1);
-//		testList.add(test2);
-//		
-//		tests.setUserTests(testList);
-		
 		
 		return tests;
 	}
